LayoutTests/requestidlecallback/requestidlecallback-is-called-expected.txt

This tests that when requestIdleCallback is enabled, requestIdleCallback is eventaully called in the order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestIdleCallbackIsCalled is true
PASS logsA.length is 2
PASS logsA.join(", ") is "1.A1, 4.A2"
PASS logsB.length is 2
PASS logsB.join(", ") is "2.B1, 3.B2"
PASS successfullyParsed is true

TEST COMPLETE
